Update to creatingThePuzzle
Instead of picking a random number every time, it now evenly distributes the amount of each number to placed into the puzzle in an array. This will decrease the chance of a number not being able to be placed each time as numbers will have a limit on how many times they can be added.